refactor: fold analyze.py into status --health, delete redundant command (Issue #1541)
Move the useful unique parts of `emdx maintain analyze` into `emdx status --health`:
- Health score calculation with metric breakdown (tag coverage, duplicate ratio,
  organization, activity)
- Tag coverage summary with untagged document count
- Actionable recommendations
- JSON output support via `--health --json`

The duplicate/similar analysis was redundant with `emdx maintain compact --dry-run`,
so users are pointed there instead.

Changes:
- Add --health/-H flag to status command with rich table output
- Add _show_health(), _show_tag_summary(), _collect_health_json() to status.py
- Remove analyze command registration from maintain.py
- Delete emdx/commands/analyze.py
- Remove 12 analyze-only TypedDicts from commands/types.py (keep HealthData,
  HealthMetricData for status --health --json)
- Update maintain docstring to reference `emdx status --health`
